requirements.txt
requests
pillow
sixel
